code/requirements.txt
mysql-connector-python==8.1.0
matplotlib==3.7.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.11.1